moved evidence enum tooltip lookup to utility function
